requirements.txt
# Web Framework
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Database
motor>=3.3.1
pymongo>=4.5.0

# Authentication and Security
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0

# File Processing
python-docx>=1.0.1
pdfplumber>=0.10.2

# LLM and AI
google-generativeai>=0.3.0

# Cache and Rate Limiting
redis>=5.0.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
uuid>=1.30